pyproject.toml
[tool.poetry]
name = "elevenlabs-mcp"
version = "0.1.0"
description = "ElevenLabs Text-to-Speech MCP Integration for Cursor"
authors = ["Sebastian Georgi"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
mcp = "^1.3.0"
fastapi = ">=0.111.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
elevenlabs = "^0.2.27"
websockets = "^12.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
aiohttp = "^3.11.13"
starlette = ">=0.41.3"
sse-starlette = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pytest-asyncio = "^0.23.5"
black = "^24.2.0"
isort = "^5.13.2"
mypy = "^1.8.0"
ruff = "^0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
elevenlabs-mcp = "src.mcp_binary.main:main"
start = "src.backend.__main__:main"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py311"
